requirements.txt
anthropic>=0.40.0
streamlit>=1.38.0
python-dotenv>=1.0.0
ddgs>=9.0.0
rich>=13.0.0
chromadb>=0.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
ruff>=0.8.0